--- a/operator/builtin/input/file/benchmark_test.go
+++ b/operator/builtin/input/file/benchmark_test.go
@@ -104,7 +104,7 @@ func BenchmarkFileInput(b *testing.B) {
 				cfg.Include = []string{
 					"file*.log",
 				}
-				cfg.MaxConcurrentFiles = 1
+				cfg.MaxConcurrentFiles = 2
 				return cfg
 			},
 		},