fixes #85 
implementing https://github.com/open-telemetry/opentelemetry-log-collection/issues/85#issuecomment-852280642

in addition 
* i removed unnecessary fingerprint comparison for copytrunc cases because it is repetitive. 
* I modified file batching test case (2nd part of test) because with this new approach, the way it handles already-discovered-recently-appended log files differently.