@rockb1017 2021-06-07
I added closing all readers step to `Stop()` function. and removed fingerprint comparison again because it is not needed.