@rockb1017 2021-06-07
hmm for windows, it fails `TestMultiFileParallel_LiveFiles` test. it consumed file1 and file4 twice. 
I can't reproduce it. any suggestion? 
Could you rerun just in case?