@rockb1017 2021-06-09
i figured out why. for MaxConcurrent, it was using 1. 
`cfg.MaxConcurrentFiles = 1`
now per batch, we open half of this number. so it wasn't able to open any file. I changed it to 2 and it works.
on this branch
```
pkg: github.com/open-telemetry/opentelemetry-log-collection/operator/builtin/input/file
BenchmarkFileInput/Single-16              504951              2003 ns/op             688 B/op         10 allocs/op
BenchmarkFileInput/Glob-16                122710              9888 ns/op            2761 B/op         42 allocs/op
BenchmarkFileInput/MultiGlob-16           123349              9809 ns/op            2761 B/op         42 allocs/op
BenchmarkFileInput/MaxConcurrent-16        65827             15271 ns/op            2757 B/op         42 allocs/op
BenchmarkFileInput/FngrPrntLarge-16       511382              2378 ns/op             690 B/op         10 allocs/op
BenchmarkFileInput/FngrPrntSmall-16       510836              2095 ns/op             688 B/op         10 allocs/op
```

on main
```
pkg: github.com/open-telemetry/opentelemetry-log-collection/operator/builtin/input/file
BenchmarkFileInput/Single-16              472940              2210 ns/op             688 B/op         10 allocs/op
BenchmarkFileInput/Glob-16                116560             12118 ns/op            2763 B/op         42 allocs/op
BenchmarkFileInput/MultiGlob-16           119990             13379 ns/op            2765 B/op         42 allocs/op
BenchmarkFileInput/MaxConcurrent-16        32445             30870 ns/op            2759 B/op         42 allocs/op
BenchmarkFileInput/FngrPrntLarge-16       508389              2413 ns/op             689 B/op         10 allocs/op
BenchmarkFileInput/FngrPrntSmall-16       474009              2588 ns/op             688 B/op         10 allocs/op

```